we fixed github

+++ lib/cli.rb
@@ -26,6 +26,8 @@ class CommandLineInterface
           menu.choice "Report a sighting", -> { self.report_sighting }
           menu.choice "See my sightings", -> { self.see_my_sightings }
           menu.choice "See my birds", -> { self.see_my_birds }
+          menu.choice "What birds are out in today's weather", -> { self.weather_choice }
+          menu.choice "Birds by location", -> { self.bird_location }
           menu.choice "Update name", -> { self.update_name }
           menu.choice "Delete a sighting", -> { self.delete_sighting }
     
@@ -60,15 +62,15 @@ class CommandLineInterface
             bird.common_name
         end
         puts bird_name_array
+        if bird_name_array == []
+            puts "There are no birds to see! Sad! So many egrets!!"
+        end
         puts "\n"
         puts " ============================== "
         prompt.select ("Back to the main menu") do |menu|
             menu.choice "Main Menu", -> { self.main_menu }
         end
-        # binding.pry
     end
-
-
     
     def see_my_sightings
         birdwatcher_sightings = self.birdwatcher.sightings.each do |sighting|
@@ -79,10 +81,40 @@ class CommandLineInterface
             sleep 2
             self.main_menu
         else
-        p birdwatcher_sightings
+            puts "\n"
+            puts " ============================== "
+            prompt.select ("Back to the main menu") do |menu|
+                menu.choice "Main Menu", -> { self.main_menu }
+            end
+        end
     end
-   
-end
+
+    def weather_choice
+        puts "What's the weather today?"
+        weather_today = gets.chomp
+        weather_choice_array = self.birdwatcher.sightings.select do |sighting|
+          sighting.weather == weather_today
+        end
+        weather_choice_array.map do |sighting|
+            puts sighting.bird.common_name
+        end
+        puts "\n"
+        puts " ============================== "
+        prompt.select ("Back to the main menu") do |menu|
+            menu.choice "Main Menu", -> { self.main_menu }
+        end
+    end
+
+    def bird_location
+        # puts "Where are you?"
+        # birdwatcher_location = gets.chomp
+        Sighting.see_all_birds_by_everyone_by_location
+        puts "\n"
+        puts " ============================== "
+        prompt.select ("Back to the main menu") do |menu|
+            menu.choice "Main Menu", -> { self.main_menu }
+        end
+    end
 
     def update_name
         puts "New identity? Put it here!"
@@ -92,13 +124,24 @@ end
     end
 
     def delete_sighting
-        puts "What do you want to delete?"
-        p see_my_sightings
-        puts "Pick a sighting's number to delete"
-        sighting_to_destroy = gets.chomp.to_i
-        byebye = Sighting.find(sighting_to_destroy)
-        byebye.destroy
-        self.main_menu
+        all_sightings = self.birdwatcher.sightings.each do |sighting|
+            puts sighting.nice_description
+        end
+        if all_sightings == []
+            puts "No sightings to delete!"
+            sleep 1
+            self.main_menu
+        else
+            puts "Pick the sighting number that you would like to delete"
+            sighting_to_destroy = gets.chomp.to_i
+            byebye = Sighting.find(sighting_to_destroy)
+            byebye.destroy
+            puts "\n"
+            puts " ============================== "
+            prompt.select ("Back to the main menu") do |menu|
+                menu.choice "Main Menu", -> { self.main_menu }
+            end
+        end
     end
 
     def goodbye
